+++ src/main/groovy/com/bmuschko/gradle/java2html/Java2HTMLPluginExtension.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 the original author or authors.
+ * Copyright 2011 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.java2html.internal.model
+package com.bmuschko.gradle.java2html
 
-import org.gradle.api.file.FileCollection
+import org.gradle.util.ConfigureUtil
 
 /**
- * Document generation input model.
+ * Java2HTML plugin extension.
  *
  * @author Benjamin Muschko
  */
-class DocGenerationInput {
-    FileCollection srcDirs
-    File destDir
-    String pattern
-    String windowTitle
-    String docTitle
-    String docDescription
-    File icon
-    File stylesheet
+class Java2HTMLPluginExtension {
+    ConversionConvention conversion = new ConversionConvention()
+    OverviewConvention overview = new OverviewConvention()
+
+    def conversion(Closure closure) {
+        ConfigureUtil.configure(closure, conversion)
+    }
+
+    def overview(Closure closure) {
+        ConfigureUtil.configure(closure, overview)
+    }
 }