+++ src/main/groovy/com/bmuschko/gradle/java2html/internal/DocMetaDataExtractor.groovy
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.java2html.internal
+package com.bmuschko.gradle.java2html.internal
 
+import com.bmuschko.gradle.java2html.internal.model.DocClass
+import com.bmuschko.gradle.java2html.internal.model.DocPackage
 import org.gradle.api.file.FileCollection
-import org.gradle.api.plugins.java2html.internal.model.DocClass
-import org.gradle.api.plugins.java2html.internal.model.DocPackage
 
 /**
  * Document meta-data extractor.