+++ src/main/groovy/com/bmuschko/gradle/java2html/internal/FileCopier.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.java2html.internal
+package com.bmuschko.gradle.java2html.internal
 
 /**
  * File copier.