+++ src/main/groovy/com/bmuschko/gradle/java2html/internal/model/DocClass.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2012 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.java2html
-
-import org.gradle.api.file.FileCollection
+package com.bmuschko.gradle.java2html.internal.model
 
 /**
- * Overview convention.
+ * Document class model.
  *
  * @author Benjamin Muschko
  */
-class OverviewConvention {
-    FileCollection srcDirs
-    File destDir
-    String pattern
-    String windowTitle
-    String docTitle
-    String docDescription
-    File icon
-    File stylesheet
+class DocClass implements Comparable {
+    String link
+    String name
+
+    @Override
+    int compareTo(t) {
+        return this.name.compareTo(t.name)
+    }
+
+    @Override
+    String toString() {
+        "DocClass{" +
+                "link='" + link + '\'' +
+                ", name='" + name + '\'' +
+                '}'
+    }
 }